requirements.txt
asgiref==3.7.2
crispy-daisyui==0.2.0
Django==5.0.2
django-appconf==1.0.6
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-imagekit==5.0.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-phonenumber-field==7.3.0
django-taggit==5.0.1
djangorestframework==3.15.1
phonenumbers==8.13.35
pilkit==3.0
pillow==10.2.0
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2024.1
